AArch64: Fix costCoeffNxN test on Apple Silicon
The assembly routine x265_costCoeffNxN_neon is buggy and produces an
incorrect result on Apple Silicon, causing the pixel testbench to fail
on these platforms.

x265_costCoeffNxN assumes that parameter `int subPosBase`, the second
parameter of type `int` passed on the stack, is at position `sp + 8`;
this assumption is consistent with the AArch64 PCS, as arguments smaller
than 8 bytes are widened to 8 bytes (aapcs64 6.8.2 C.16).
However arm64e diverges from AAPCS64: 'Function arguments may consume
slots on the stack that are not multiples of 8 bytes'.