Test: Relax constraints of check_saoCuStatsE*
The tests `check_saoCuStatsE*` check for equality of auxiliary
buffers used to store intermediate results for re-use between rows
(upBuff*). This constraint is not strictly necessary, as the buffers do
not form part of the result of `saoCuStatsE*` and are re-initialized
before every call to the routine.

This patch removes the equality test of the buffer(s), to allow certain
optimisations to be made in SIMD primitives without causing the tests to
fail.