AArch64: Build with -flax-vector-conversions=none
Currently -flax-vector-conversions is passed to GCC (and all compilers
that are treated as GCC-like in source/CMakeLists.txt) when building for
AArch64.

Remove this, and add -flax-vector-conversions=none for compilers which
support this, to enforce explicit vector type conversions. GCC already
does not allow implicit vector type conversions by default.