AArch64: Fix implicit conversion from ptrdiff_t to int
In common/aarch64/loopfilter-prim.cpp there is a function call to abs()
with the result of subtracting two pointer types, which results in an
implicit cast from ptrdiff_t to int as abs() is only defined for int
prior to C++11 (and x265 uses gnu++98).

Make conversion from ptrdiff_t to int explicit by using a cast in
common/aarch64/loopfilter-prim.cpp.