AArch64: Optimise Neon assembly implementations of SADxN
Optimise the Neon assembly implementations of SADx3 and SADx4
primitives, replacing UABAL, UABAL2 sequences with UABD, UADALP which
has twice the throughput on modern Arm cores.

Also refactor the load instructions for block sizes of width 4 to use
LDR for the first partial load of a vector register - making the
operation completely destructive.

As this patch refactors some of the block sizes (12xh, 16xh) to use the
LOOP macro (rather than the fully unrolled macro), the SVE2
implementations which make use of these Neon macros are updated as
required.