AArch64: Add check for arm_neon_sve_bridge.h
In order to enable future additions of SVE/SVE2 intrinsics code, add a
test to the CMake configuration to check for the availability of the
arm_neon_sve_bridge.h header.

To simplify this check, and tests for AArch64 architecture features in
general, set all CPU_HAS_<FEAT> flag(s) that are available for a
particular architecture feature, including when cross compiling. Also
clean up some unnecessary comments.