AArch64: Clean up CMake feature detection
Neon is a compulsory ISA feature for AArch64, so remove test for Neon
when compiling for AArch64. SVE2 is an extension from Armv9.0-A, so use
the armv9-a architecture flag when compiling with SVE2.

Also clean up some duplicated compiler flags so that these are added in
one place, and handle cross-compilation options for feature extensions
separately. Add SVE and SVE2 assembly sources in separate, non-exclusive
conditional blocks to remove duplication, and to simplify the future
addition of new architecture features. Refactor these conditional checks
to test for the CPU_HAS_<FEAT> flag only - as this is now also set when
cross-compiling. Similarly, remove the check for CROSS_COMPILE_ARM64, as
the variable ARM64 is set via checking CMAKE_SYSTEM_PROCESSOR when cross
compiling.