Test: Update values used in check_nquant_primitive
Change the calculation of the `qBits` parameter in
check_nquant_primitive to be more similar to the calling code in
quant.cpp, rather than using `rand() % 32` to generate a value. The
changes are based on the generation of qBits in check_quant_primitive.

This ensures that the `qBits` parameter in the test is always non-zero,
in order to allow use of non-saturating narrowing instructions in
optimised nquant implementations.